Reworked data handling.
The data will now be read continously and the transmission in each direction will be delayed. This should make it so that the delay is percieved as latency and not just a slowing down of the passthrough.

Data is buffered in memory until it needs to be sent. As long as the output bandwidth is equal to or greater than the input bandwidth this should accurately simulate network latency.